Reorganize
MedPC loader now similar to other loaders. Maze_breakdown deleted
and functions moved to utils or deleted if using matlab.
