+++ nept/__init__.py
@@ -8,14 +8,13 @@ from .co_occurrence import spike_counts, get_tetrode_mask, compute_cooccur
 from .decoding import bayesian_prob, decode_location, remove_teleports
 from .lfp_filtering import detect_swr_hilbert
 from .loaders_mclust import load_mclust_header, load_spikes
+from .loaders_medpc import load_medpc
 from .loaders_neuralynx import (load_events,
                                 load_lfp,
                                 load_position,
                                 load_ntt,
                                 load_neuralynx_header,
                                 load_nvt)
-from .maze_breakdown import expand_line, save_spike_position
-from .medpc import load_medpc
 from .place_fields import (consecutive,
                            find_fields,
                            get_single_field,
@@ -25,6 +24,7 @@ from .tuning_curves import (binned_position,
                             tuning_curve_2d)
 from .utils import (bin_spikes,
                     cartesian,
+                    expand_line,
                     find_multi_in_epochs,
                     find_nearest_idx,
                     find_nearest_indices,
