+++ nept/lfp_filtering.py
@@ -1,7 +1,6 @@
 import numpy as np
 import scipy.signal
 import scipy.stats as stats
-
 import nept
